+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Ignored hot key command on edit fields. Refs UIIN-2604.
 * Don't render Fast Add record modal in a `<Paneset>` to re-calculate other `<Pane>`'s widths after closing. Fixes UIIN-2690.
 * Don't include `sort` parameter in requests to facets. Fixes UIIN-2685.
+* Inventory app > Search and Browse Results > Update HTML page title with search term entered. Refs UIIN-2581.
 
 ## [10.0.4](https://github.com/folio-org/ui-inventory/tree/v10.0.4) (2023-11-10)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v10.0.3...v10.0.4)